notebooks/helper_scripts/requirements.txt
python-dotenv
numpy
sentencepiece